refactor: Implement naming change for type defintions
