+++ plugin-hrm-form/src/states/configuration/hooks/useProfileFlagDurations.ts
